+++ addons/monitoring/monitoring-core/monitoring-core-common/pom.xml
@@ -29,10 +29,6 @@
       <groupId>org.kie.kogito</groupId>
       <artifactId>grafana-api</artifactId>
     </dependency>
-    <dependency>
-      <groupId>org.kie.kogito</groupId>
-      <artifactId>explainability-addon-common</artifactId>
-    </dependency>
 
     <dependency>
       <groupId>io.micrometer</groupId>
